+++ bitreq/src/proxy.rs
@@ -18,7 +18,7 @@ pub(crate) enum ProxyKind {
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub struct Proxy {
     pub(crate) server: String,
-    pub(crate) port: u32,
+    pub(crate) port: u16,
     pub(crate) user: Option<String>,
     pub(crate) password: Option<String>,
     pub(crate) kind: ProxyKind,
@@ -33,9 +33,9 @@ impl Proxy {
         }
     }
 
-    fn parse_address(host: &str) -> Result<(String, Option<u32>), Error> {
+    fn parse_address(host: &str) -> Result<(String, Option<u16>), Error> {
         if let Some((host, port)) = split_once(host, ":") {
-            let port = port.parse::<u32>().map_err(|_| Error::BadProxy)?;
+            let port = port.parse::<u16>().map_err(|_| Error::BadProxy)?;
             Ok((host.to_string(), Some(port)))
         } else {
             Ok((host.to_string(), None))
